Provide configurability for Valiant
This is a fit-out of the configuration side of Valiant that includes
an overlay approach to loading config sources. The broad changes
indicate the refinement of the configuration aspect.

Note that I have removed the report provider configuration code
as this will be reviewed as I setup a plugin model.

Issues: #4 and #5