Implementation of plugins
The valiant.plugins package contains the implementation.

I've shuffled the report providers (basic, safety & spdx) into
the plugins.reports package as they now operate as plugins.

The plugin work involved further modifications to the
config-related code. This allows for local plugins and is
also a tidy up of the map->config approach.

The tests have been reconfigured as needed but a range
of new tests are required.

Issues: #4 #5 #6